requirements.txt
Flask
flask-cors
sentence-transformers
faiss-cpu
requests
mysql-connector-python